+++ common/src/main/java/net/mehvahdjukaar/supplementaries/SuppPlatformStuff.java
@@ -1,6 +1,7 @@
 package net.mehvahdjukaar.supplementaries;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
+import net.mehvahdjukaar.moonlight.api.fluids.SoftFluidStack;
 import net.mehvahdjukaar.supplementaries.common.utils.SlotReference;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;